Update OperacionesMatematicas.java
Changed the return type for addition, substraction and multiplication

+++ src/Repaso/OperacionesMatematicas.java
@@ -1,13 +1,11 @@
 package Repaso;
 
 public interface OperacionesMatematicas {
-    public double sumar(int a, int b);
+    public int sumar(int a, int b);
 
-    public double restar(int a, int b);
+    public int restar(int a, int b);
 
-    public double multiplicar(int a, int b);
+    public int multiplicar(int a, int b);
 
     public double dividir(int a, int b);
-
-    // addition, subtraction, multiplication and division
 }
